Specify the Puppet confdir in Bolt apply
Bolt overwrites Puppet settings to ensure Bolt apply and puppet agent
runs do not conflict, or write to the same log files, etc. So we look up
the real confdir with a command and specify the confdir parameters so
that puppetdb::master::config writes the ini settings to the proper
files.

Because Bolt's Puppet doesn't have the config settings of the server's
Puppet this means that the Puppet used in Bolt apply doesn't have the settings
necessary for the PuppetDB module to check a valid connection between
Puppetserver and PuppetDB, so we must to disable `strict_validation`.