Add ability to create agents and autosign certs
Adds the ability to specify an autosign whitelist for your puppetserver
to the bolt dev::foss_master plan. It specifies a list of FQDN globs to
whitelist and will be written to the contents of
`/etc/puppetlabs/puppet/autosign.conf`.

To the vmpooler script, this commit the ability to pull extra nodes from
vmpooler (of the same platform as the master) and connect them to the
puppetmaster. This assumes they can be autosigned by configuring the
autosign whitelist.