Open port 8140 on master for puppetserver
By default the `manage_firewall` param is `true` for the `puppetdb`
module, and it rejects connections for all ports except 22 (for ssh).
This prevents puppet agents from checking in with puppetserver to get a
signed certiifcate, so we need to open port 8140 for connections.

So far, there's no need to open PuppetDB's ports, but if so the
`open_http_port` and `open_ssl_port` parameters exist for that.