Install master via `puppet agent` run
This switches the puppet master from using `puppet apply` with a heredoc
to installing a puppet master via a regular `puppet agent` catalog
compilation.

There's a bit of a chicken and an egg problem with doing it this way so
there's still a little bit of leg work necessary to setup puppetserver
and put the necessary code in the codedir before the manifest can be
applied. But once setup, the manifest installs and configures PuppetDB,
PostgreSQL, a Puppetserver -> PuppetDB connection, report storage, the
autosign whitelist, and the firewall to allow agents to connect to the
puppetserver