Add --modules flag to install default modules
The modules flag can be specified to install a default set of comma
separated modules and include them for all nodes (such as ntp,
os_patching, etc)

Example:

```
./bin/vmpooler/foss-master --modules albatrossflavour-os_patching,puppetlabs-ntp
```

This will use the `puppet module install` command to install both
modules on the master and `include os_patching` and `include ntp` on all
nodes. Normally this would be handled by r10k and the manifests in a
control repo, but none of that has been configured for this dev install.